ci: drop conjugator for now
