+++ .github/workflows/tests.yml
@@ -65,7 +65,7 @@ jobs:
         if: matrix.python-version != '3.10' && matrix.python-version != '3.12' && matrix.python-version != '3.13' && matrix.python-version != '3.14'
 
       - name: Install dependencies
-        run: uv pip install -e . --group dev --group setup
+        run: uv pip install -e . --group dev
         if: matrix.python-version == '3.10'
 
       - name: Install dependencies
